Refactor PackageRestoreDataSource
