+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/PackageRestore/PackageRestoreDataSource.cs
@@ -22,7 +22,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore
     [Export(typeof(IPackageRestoreDataSource))]
     [Export(ExportContractNames.Scopes.UnconfiguredProject, typeof(IProjectDynamicLoadComponent))]
     [AppliesTo(ProjectCapability.PackageReferences)]
-    internal class PackageRestoreDataSource : ChainedProjectValueDataSourceBase<RestoreData>, IPackageRestoreDataSource, IProjectDynamicLoadComponent, IVsProjectRestoreInfoSource
+    internal partial class PackageRestoreDataSource : ChainedProjectValueDataSourceBase<RestoreData>, IPackageRestoreDataSource, IProjectDynamicLoadComponent
     {
         // This class represents the last data source in the package restore chain, which is made up of the following:
         //  _________________________________________      _________________________________________
@@ -66,28 +66,6 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore
         private byte[]? _latestHash;
         private bool _enabled;
 
-        /// <summary>
-        /// Re-usable task that completes when there is a new nomination
-        /// </summary>
-        private TaskCompletionSource<bool>? _whenNominatedTask;
-
-        private bool _restoreStarted;
-
-        private bool _wasSourceBlockContinuationSet;
-
-        /// <summary>
-        /// Save the configured project versions that might get nominations.
-        /// </summary>
-        private readonly Dictionary<ProjectConfiguration, IComparable> _savedNominatedConfiguredVersion = new();
-
-        /// <summary>
-        /// Project Unique Name used by Nuget Nomination.
-        /// </summary>
-        public string Name => _project.FullPath;
-
-        // True means the project system plans to call NominateProjectAsync in the future.
-        bool IVsProjectRestoreInfoSource.HasPendingNomination => CheckIfHasPendingNomination();
-
         [ImportingConstructor]
         public PackageRestoreDataSource(
             UnconfiguredProject project,
@@ -113,23 +91,7 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore
 
         protected override IDisposable? LinkExternalInput(ITargetBlock<IProjectVersionedValue<RestoreData>> targetBlock)
         {
-            // Register before this project receives any data flows containing possible nominations.
-            // This is needed because we need to register before any nuget restore or before the solution load.
-#pragma warning disable RS0030 // Do not used banned APIs
-            var registerRestoreInfoSourceTask = Task.Run(async () =>
-            {
-                try
-                {
-                    await _solutionRestoreService4.RegisterRestoreInfoSourceAsync(this, _projectAsynchronousTasksService.UnloadCancellationToken);
-                }
-                catch (Exception ex)
-                {
-                    throw ex;
-                }
-            });
-#pragma warning restore RS0030 // Do not used banned APIs
-
-            _project.Services.FaultHandler.Forget(registerRestoreInfoSourceTask, _project, ProjectFaultSeverity.Recoverable);
+            RegisterProjectRestoreInfoSource();
 
             JoinUpstreamDataSources(_dataSource);
 
@@ -212,27 +174,6 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore
             return success;
         }
 
-        private void SaveNominatedConfiguredVersions(IReadOnlyCollection<PackageRestoreConfiguredInput> configuredInputs)
-        {
-            lock (SyncObject)
-            {
-                _savedNominatedConfiguredVersion.Clear();
-
-                foreach (var configuredInput in configuredInputs)
-                {
-                    _savedNominatedConfiguredVersion[configuredInput.ProjectConfiguration] = configuredInput.ConfiguredProjectVersion;
-                }
-
-                if (_whenNominatedTask is not null)
-                {
-                    if (_whenNominatedTask.TrySetResult(true))
-                    {
-                        _whenNominatedTask = null;
-                    }
-                }
-            }
-        }
-
         private void HintProjectDependentFile(ProjectRestoreInfo restoreInfo)
         {
             if (restoreInfo.ProjectAssetsFilePath.Length != 0)
@@ -300,164 +241,6 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore
             }
 
             return Task.CompletedTask;
-        }
-
-        // NuGet calls this method to wait project to nominate restoring.
-        // If the project has no pending restore data, it will return a completed task.
-        // Otherwise a task which will be completed once the project nominate the next restore
-        // the task will be cancelled, if the project system decide it no longer need restore (for example: the restore state has no change)
-        // the task will be failed, if the project system runs into a problem, so it cannot get correct data to nominate a restore (DT build failed)
-        public Task WhenNominated(CancellationToken cancellationToken)
-        {
-            lock (SyncObject)
-            {
-                if (cancellationToken.IsCancellationRequested)
-                {
-                    cancellationToken.ThrowIfCancellationRequested();
-                }
-
-                if (!CheckIfHasPendingNomination())
-                {
-                    return Task.CompletedTask;
-                }
-
-                _whenNominatedTask ??= new TaskCompletionSource<bool>(TaskCreationOptions.RunContinuationsAsynchronously);
-
-                if (!_wasSourceBlockContinuationSet)
-                {
-                    _wasSourceBlockContinuationSet = true;
-
-                    _ = SourceBlock.Completion.ContinueWith(t =>
-                    {
-                        lock (SyncObject)
-                        {
-                            if (_whenNominatedTask is not null)
-                            {
-                                if (t.IsFaulted)
-                                {
-                                    _whenNominatedTask.SetException(t.Exception);
-                                }
-                                else
-                                {
-                                    _whenNominatedTask.TrySetCanceled();
-                                }
-                            }
-                        }
-                    }, CancellationToken.None, TaskContinuationOptions.OnlyOnFaulted, TaskScheduler.Default);
-                }
-            }
-
-            return _whenNominatedTask.Task.WithCancellation(cancellationToken);
-        }
-
-        private bool CheckIfHasPendingNomination()
-        {
-            lock (SyncObject)
-            {
-                Assumes.Present(_project.Services.ActiveConfiguredProjectProvider);
-                Assumes.Present(_project.Services.ActiveConfiguredProjectProvider.ActiveConfiguredProject);
-
-                // Nuget should not wait for projects that failed DTB
-                if (!_enabled || SourceBlock.Completion.IsFaulted || SourceBlock.Completion.IsCompleted)
-                {
-                    return false;
-                }
-
-                // Avoid possible deadlock.
-                // Because RestoreCoreAsync() is called inside a dataflow block it will not be called with new data
-                // until the old task finishes. So, if the project gets nominating restore, it will not get updated data.
-                if (IsPackageRestoreOnGoing())
-                {
-                    return false;
-                }
-
-                ConfiguredProject? activeConfiguredProject = _project.Services.ActiveConfiguredProjectProvider.ActiveConfiguredProject;
-
-                // After the first nomination, we should check the saved nominated version
-                return IsSavedNominationOutOfDate(activeConfiguredProject);
-            }
-        }
-
-        private bool IsPackageRestoreOnGoing()
-        {
-            // If NominateForRestoreAsync() has not finished, return false
-            return _restoreStarted;
-        }
-
-        protected virtual bool IsSavedNominationOutOfDate(ConfiguredProject activeConfiguredProject)
-        {
-            if (!_savedNominatedConfiguredVersion.TryGetValue(activeConfiguredProject.ProjectConfiguration,
-                    out IComparable latestSavedVersionForActiveConfiguredProject) ||
-                activeConfiguredProject.ProjectVersion.IsLaterThan(latestSavedVersionForActiveConfiguredProject))
-            {
-                return true;
-            }
-
-            if (_savedNominatedConfiguredVersion.Count == 1)
-            {
-                return false;
-            }
-
-            foreach (var loadedProject in activeConfiguredProject.UnconfiguredProject.LoadedConfiguredProjects)
-            {
-                if (_savedNominatedConfiguredVersion.TryGetValue(loadedProject.ProjectConfiguration, out IComparable savedProjectVersion))
-                {
-                    if (loadedProject.ProjectVersion.IsLaterThan(savedProjectVersion))
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        protected virtual bool IsProjectConfigurationVersionOutOfDate(IReadOnlyCollection<PackageRestoreConfiguredInput> configuredInputs)
-        {
-            Assumes.Present(_project.Services.ActiveConfiguredProjectProvider);
-            Assumes.Present(_project.Services.ActiveConfiguredProjectProvider.ActiveConfiguredProject);
-
-            if (configuredInputs is null)
-            {
-                return false;
-            }
-
-            var activeConfiguredProject = _project.Services.ActiveConfiguredProjectProvider.ActiveConfiguredProject;
-
-            IComparable? activeProjectConfigurationVersionFromConfiguredInputs = null;
-            foreach (var configuredInput in configuredInputs)
-            {
-                if (configuredInput.ProjectConfiguration.Equals(activeConfiguredProject.ProjectConfiguration))
-                {
-                    activeProjectConfigurationVersionFromConfiguredInputs = configuredInput.ConfiguredProjectVersion;
-                }
-            }
-
-            if (activeProjectConfigurationVersionFromConfiguredInputs is null ||
-                activeConfiguredProject.ProjectVersion.IsLaterThan(
-                    activeProjectConfigurationVersionFromConfiguredInputs))
-            {
-                return true;
-            }
-
-            if (configuredInputs.Count == 1)
-            {
-                return false;
-            }
-
-            foreach (var loadedProject in activeConfiguredProject.UnconfiguredProject.LoadedConfiguredProjects)
-            {
-                foreach (var configuredInput in configuredInputs)
-                {
-                    if (loadedProject.ProjectConfiguration.Equals(configuredInput.ProjectConfiguration) &&
-                        loadedProject.ProjectVersion.IsLaterThan(configuredInput.ConfiguredProjectVersion))
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
         }
     }
 }
